Refactor sequential runner tests into scenario modules
