+++ projects/04-llm-adapter-shadow/tests/sequential/__init__.py
@@ -1,0 +1,1 @@
+"""Sequential runner test suite."""
